steps as ints all day

+++ orix/utils/grid_utils/basic_grid_utils.py
@@ -36,7 +36,7 @@ def create_linearly_spaced_array_in_szxz(resolution):
 
     num_steps = int(360/resolution + 0.5)
     alpha = np.linspace(0,360,num=num_steps,endpoint=False)
-    beta  = np.linspace(0,180,num=num_steps/2,endpoint=False)
+    beta  = np.linspace(0,180,num=int(num_steps/2),endpoint=False)
     gamma = np.linspace(0,360,num=num_steps,endpoint=False)
     z = np.asarray(list(product(alpha, beta, gamma)))
     return Euler(z,axis_convention='szxz')
